Reopen Component and View for 1.13+ compat.
 Please enter the commit message for your changes. Lines starting
