+++ addon/overrides/checkbox.js
@@ -1,0 +1,8 @@
+import Ember from "ember";
+import MobiletouchMixin from './mobiletouch-mixin';
+
+// Ember.Checkbox in 1.13 extends from the private/internal `Ember.View` that
+// we cannot reopen in a 'normal' fashion, this reopens it directly
+//
+// In Ember 2.0.0 Ember.Checkbox will extend from Ember.Component
+export default Ember.Checkbox.reopen(MobiletouchMixin);
